pyproject.toml
[project]
name = "tool-use-ft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.4",
    "openai>=1.95.0",
    "pandas>=2.3.1",
    "peft>=0.16.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "together>=1.5.21",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.1",
    "trl>=0.19.1",
]
